requirements.txt
Flask==2.3.3
gtts==2.5.4
google-generativeai==0.3.2 